+++ test-packages/test-utils/api.d.ts
@@ -15,9 +15,10 @@ import {
   UI5Field,
   UI5EnumValue
 } from "@ui5-language-assistant/semantic-model-types";
-import { Json, TypeNameFix } from "@ui5-language-assistant/semantic-model";
+import { TypeNameFix } from "@ui5-language-assistant/semantic-model";
 import { XMLAttribute, XMLElement } from "@xml-tools/ast";
 import { UI5XMLViewCompletion } from "@ui5-language-assistant/xml-views-completion";
+import { FetchResponse } from "@ui5-language-assistant/language-server";
 
 //	easily build (partial) data structures for tests with mandatory "name" field
 export type PartialWithName<T> = { name: string } & Partial<T>;
@@ -68,19 +69,25 @@ export function buildUI5Model(
   opts: Partial<UI5SemanticModel>
 ): UI5SemanticModel;
 
-export type TestModelVersion = "1.60.14" | "1.74.0" | "1.75.0";
+export type TestModelVersion = "1.60.14" | "1.74.0" | "1.75.0" | "1.71.14";
 
 /**
  * Return a UI5SemanticModel for the specified version.
- * @param version
- * @param disableCache By default caching is used. This improves performance and is safe (since the model is immutable).
+ *
+ * If downloadLibraries is true (default), increase the timeout of the test/hook to GEN_MODEL_TIMEOUT.
+ *
+ * @param opts.version
+ * @param opts.downloadLibraries By default, download the library files before creating the model.
+ *        If you call downloadLibraries explicitly before this function, send false (it will improve the performance).
+ * @param opts.fixLibs Apply fixes to the model so that it can be created in strict mode. True by default.
+ * @param opts.strict Generate the model in strict mode. True by default.
  */
-export function generateModel(
-  version: TestModelVersion,
-  disableCache?: boolean
-): UI5SemanticModel;
-
-export function loadLibraries(version: TestModelVersion): Record<string, Json>;
+export function generateModel(opts: {
+  version: TestModelVersion;
+  downloadLibs?: boolean;
+  fixLibs?: boolean;
+  strict?: boolean;
+}): Promise<UI5SemanticModel>;
 
 export function getTypeNameFixForVersion(
   version: TestModelVersion
@@ -119,3 +126,17 @@ export function expectSuggestions(
   suggestions: UI5XMLViewCompletion[],
   expected: string[]
 ): void;
+
+export function readTestLibraryFile(
+  version: string,
+  fileName: string
+): Promise<FetchResponse>;
+
+export function downloadLibraries(version: TestModelVersion): Promise<void>;
+
+/**
+ * generateModel can take a long time in some cases now that it might download the libraries.
+ * Use this constant to set the timeout on the test/hook that calls it.
+ * Note: this.timeout(...) cannot be called from within an arrow function.
+ */
+export const GEN_MODEL_TIMEOUT = 5000;