+++ packages/logic-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/sap/ui5-language-assistant/compare/@ui5-language-assistant/logic-utils@1.1.0...@ui5-language-assistant/logic-utils@1.1.1) (2020-04-07)
+
+**Note:** Version bump only for package @ui5-language-assistant/logic-utils
+
 # [1.1.0](https://github.com/sap/ui5-language-assistant/compare/@ui5-language-assistant/logic-utils@1.0.0...@ui5-language-assistant/logic-utils@1.1.0) (2020-04-01)
 
 ### Features